move neo_compare
