+++ p8_integration_tests/quick_test/test_multi_call_examples/test_reset_different.py
@@ -14,7 +14,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import spynnaker8 as sim
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.checker import check_neuron_data
