+++ p8_integration_tests/x_test/test_0_1_time_steps/test_malloc_key_allocator/test_synfire_if_curr_exp_with_malloc_key_allocator.py
@@ -22,7 +22,7 @@ from neo.io import PickleIO
 
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 from spinnman.exceptions import SpinnmanTimeoutException
 from unittest import SkipTest
