+++ p8_integration_tests/x_test/test_0_1_time_steps/test_spikes/test_synfire_0dot1_timestep_test_print_spikes.py
@@ -23,7 +23,7 @@ from unittest import SkipTest
 from spinnman.exceptions import SpinnmanTimeoutException
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 
 n_neurons = 20
 timestep = 0.1
