+++ p8_integration_tests/x_test/test_0_1_time_steps/test_v/test_synfire_0dot1_timestep_test_print_v.py
@@ -22,7 +22,7 @@ from neo.io import PickleIO
 import unittest
 from unittest import SkipTest
 from spinnman.exceptions import SpinnmanTimeoutException
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
