+++ p8_integration_tests/x_test/test_1_0_time_steps/test_malloc_key_allocator/test_synfire_if_curr_exp_with_malloc_key_allocator.py
@@ -18,7 +18,7 @@ from neo.io import PickleIO
 import unittest
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 from spinn_front_end_common.utilities.exceptions import ConfigurationException
 
 synfire_run = SynfireRunner()
