+++ p8_integration_tests/x_test/test_1_0_time_steps/test_v/test_synfire_1dot0_timestep_test_get_v.py
@@ -22,7 +22,7 @@ from neo.io import PickleIO
 import unittest
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 
 n_neurons = 200  # number of neurons in each population
 runtime = 500
