+++ p8_integration_tests/x_test/test_1_0_time_steps/test_va_bench_mark_tests/test_va_benchmark.py
@@ -38,7 +38,7 @@ import unittest
 from unittest import SkipTest
 from p8_integration_tests.base_test_case import BaseTestCase
 import spynnaker8 as p
-from spynnaker8.utilities import neo_compare
+from spynnaker.pyNN.utilities import neo_compare
 from pyNN.random import NumpyRNG, RandomDistribution
 from pyNN.utility import Timer
