+++ spynnaker/pyNN/utilities/neo_compare.py
